+++ src/CleanRegex/Exception/FormatMalformedPatternException.php
@@ -1,6 +1,0 @@
-<?php
-namespace TRegx\CleanRegex\Exception;
-
-class FormatMalformedPatternException extends PatternMalformedPatternException
-{
-}