+++ src/CleanRegex/Exception/MaskMalformedPatternException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\CleanRegex\Exception;
+
+class MaskMalformedPatternException extends PatternMalformedPatternException
+{
+}